+++ src/History.js
@@ -51,7 +51,7 @@ define(function (require) {
             .on("click.history", ".history-commit", function () {
                 var hash = $(this).attr("x-hash");
                 var commit = _.find(commitCache, function (commit) { return commit.hash === hash; });
-                HistoryViewer.show(commit);
+                HistoryViewer.show(commit, getCurrentDocument());
             });
     }